# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "easyOrder 3.0",
    "description": "Sistema de Gestão de Pedidos - Tech Challenge (3) da Pos Tech SOAT ",
    "version": "1.0.0"
  },
  "host": "",
  "paths": {
...
    "/produto/listar": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Listar Produtos",
        "description": "Lista todos os produtos cadastrados<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/produto/buscar/{id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Buscar Produto",
        "description": "Busca um produto utilizando o Id.<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do produto",
            "example": "5e73a938-41e7-4b76-a5a0-ae6147266e72"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto localizado:",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "resultado_busca": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Produto localizado:"
                    },
                    "produto": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "5e73a938-41e7-4b76-a5a0-ae6147266e72"
                        },
                        "nome": {
                          "type": "string",
                          "example": "X-Bacon"
                        },
                        "preco": {
                          "type": "number",
                          "example": "20"
                        },
                        "categoria": {
                          "type": "string",
                          "example": "Lanche"
                        }
                      }
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "resultado_busca": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Produto localizado:"
                    },
                    "produto": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "5e73a938-41e7-4b76-a5a0-ae6147266e72"
                        },
                        "nome": {
                          "type": "string",
                          "example": "X-Bacon"
                        },
                        "preco": {
                          "type": "number",
                          "example": "20"
                        },
                        "categoria": {
                          "type": "string",
                          "example": "Lanche"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "resultado_busca": {
                      "type": "boolean",
                      "example": false
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Produto não encontrado"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "resultado_busca": {
                      "type": "boolean",
                      "example": false
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Produto não encontrado"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/produto/listar/{categoria}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Listar Produtos por Categoria",
        "description": "Lista produtos por categoria<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "categoria",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da Categoria",
            "example": "LANCHE"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/produto/remover/{id}": {
      "delete": {
        "tags": [
          "Produtos"
        ],
        "summary": "Remover Produto",
        "description": "Remove um produto, por Id.<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do produto",
            "example": "228ec10e-5675-47f4-ba1f-2c4932fe68cc"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/produto/cadastrar": {
      "post": {
        "tags": [
          "Produtos"
        ],
        "summary": "Cadastrar Produto",
        "description": "Realiza o Cadastro de um Novo Produto, através dos dados fornecidos no corpo da requisição.<br><br>  [ Endpoint para integração aos sistemas administrativo e/ou de loja ]",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "string",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "X-Salada"
                },
                "descricao": {
                  "type": "string",
                  "example": "sem salada, sem tomate"
                },
                "preco": {
                  "type": "number",
                  "example": "25"
                },
                "imagemURL": {
                  "type": "string",
                  "example": "xsalada.png"
                },
                "categoria": {
                  "type": "string",
                  "example": "LANCHE"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "resultado_cadastro": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Produto cadastrado com sucesso"
                    },
                    "produto": {
                      "type": "object",
                      "properties": {
                        "nome": {
                          "type": "string",
                          "example": "X-Salada"
                        },
                        "descricao": {
                          "type": "string",
                          "example": "sem salada, sem tomate"
                        },
                        "preco": {
                          "type": "number",
                          "example": "25"
                        },
                        "imagemURL": {
                          "type": "string",
                          "example": "xsalada.png"
                        },
                        "categoria": {
                          "type": "string",
                          "example": "LANCHE"
                        }
                      }
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "resultado_cadastro": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string",
                      "example": "Produto cadastrado com sucesso"
                    },
                    "produto": {
                      "type": "object",
                      "properties": {
                        "nome": {
                          "type": "string",
                          "example": "X-Salada"
                        },
                        "descricao": {
                          "type": "string",
                          "example": "sem salada, sem tomate"
                        },
                        "preco": {
                          "type": "number",
                          "example": "25"
                        },
                        "imagemURL": {
                          "type": "string",
                          "example": "xsalada.png"
                        },
                        "categoria": {
                          "type": "string",
                          "example": "LANCHE"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ocorreu um erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Erro inesperado: Não foi possivel cadastrar o produto"
                    }
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "properties": {
                    "mensagem": {
                      "type": "string",
                      "example": "Erro inesperado: Não foi possivel cadastrar o produto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  },
                  "descricao": {
                    "example": "any"
                  },
                  "preco": {
                    "example": "any"
                  },
                  "categoria": {
                    "example": "any"
                  },
                  "imagemURL": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}